some moving things around and refactoring
