+++ scripts/crontab
@@ -8,7 +8,7 @@ PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/us
 
 ## GITCOIN WEB3 STUFF
 ##TODO: Re-enable this when Gitcoin GETH URL is synced.
-##* * * * * cd gitcoin/coin; bash scripts/run_management_command_if_not_already_running.bash sync_geth --provider custom  >> /var/log/gitcoin/sync_geth.log  2>&1
+*/15 * * * * cd gitcoin/coin; bash scripts/run_management_command_if_not_already_running.bash sync_geth mainnet  >> /var/log/gitcoin/sync_geth.log  2>&1
 
 ## TOOLING
 1 */3 * * * cd gitcoin/coin; bash scripts/run_management_command_if_not_already_running.bash get_prices  >> /var/log/gitcoin/get_prices.log  2>&1
